+++ react/__tests__/components/SKUSelector.test.js
@@ -1,5 +1,6 @@
 import React from 'react'
 import { render } from 'intl-helper'
+import { fireEvent } from 'react-testing-library'
 
 import SKUSelector from './../../SKUSelector'
 import { getSKU } from 'sku-helper'
@@ -22,5 +23,13 @@ describe('<SKUSelector />', () => {
   it('should match the snapshot', () => {
     const { asFragment } = renderComponent()
     expect(asFragment()).toMatchSnapshot()
+  })
+
+  it('should call onSKUSelected', () => {
+    const onSKUSelected = jest.fn()
+    const { container } = renderComponent({ onSKUSelected })
+    const selector = container.querySelector('.skuSelectorItem')
+    fireEvent.click(selector)
+    expect(onSKUSelected).toBeCalledTimes(1)
   })
 })