+++ CHANGELOG.md
@@ -7,8 +7,14 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+
+- Add behavior test to SKUSelector.
+
 ## [3.19.6] - 2019-03-11
+
 ### Fixed
+
 - Change current product image when pressing on a thumbnail.
 
 ### Added